Add more confirmation dialogs

+++ src/frontend/app/shared/components/list/list-types/cf-orgs/cf-org-card/cf-org-card.component.ts
@@ -23,6 +23,8 @@ import { CardCell } from '../../../list.types';
 import { ComponentEntityMonitorConfig } from '../../../../../shared.types';
 import { getEntityDeleteSections } from '../../../../../../store/selectors/api.selectors';
 import { entityFactory, organizationSchemaKey } from '../../../../../../store/helpers/entity-factory';
+import { ConfirmationDialogService } from '../../../../confirmation-dialog.service';
+import { ConfirmationDialogConfig } from '../../../../confirmation-dialog.config';
 
 @Component({
   selector: 'app-cf-org-card',
@@ -48,7 +50,8 @@ export class CfOrgCardComponent extends CardCell<APIResource<IOrganization>> imp
     private cfUserService: CfUserService,
     private cfEndpointService: CloudFoundryEndpointService,
     private store: Store<AppState>,
-    private currentUserPermissionsService: CurrentUserPermissionsService
+    private currentUserPermissionsService: CurrentUserPermissionsService,
+    private confirmDialog: ConfirmationDialogService
   ) {
     super();
 
@@ -127,10 +130,18 @@ export class CfOrgCardComponent extends CardCell<APIResource<IOrganization>> imp
   }
 
   delete = () => {
-    this.cfEndpointService.deleteOrg(
-      this.row.metadata.guid,
-      this.cfEndpointService.cfGuid
+    const confirmation = new ConfirmationDialogConfig(
+      'Delete Organization',
+      `Are you sure you want to delete organization '${this.row.entity.name}'?`,
+      'Delete',
+      true
     );
+    this.confirmDialog.open(confirmation, () => {
+      this.cfEndpointService.deleteOrg(
+        this.row.metadata.guid,
+        this.cfEndpointService.cfGuid
+      );
+    });
   }
 
   goToSummary = () => this.store.dispatch(new RouterNav({
